Utilize positive lookbehind assertion
Using "(?<=,)|(?=\\))" in "parameter-ownership".

`(?<=,)` is a positive lookbehind assertion that matches a position in
the text only if it is immediately preceded by a comma `,`.
It does not consume any characters in the match.

`(?=\\)` is a positive lookahead assertion that matches a position in
the text only if it is immediately followed by a closing parenthesis `)`.

This allows properly including other pattern that have a match which
which ends with a comma `,`.